Teach RequireJS how to load geodesy.
geodesy is a CommonJS module, which could be loaded by RequireJS's
compatibility mode, except a bug in RequireJS is tickled by the
'require("./dms.js")' line. Instead, this loads it as a bare
JavaScript file, polluting the window object with LatLon and Dms
globals, and then injects them into the RequireJS module system.