Enable RequireJS's node module support.
Some modules, like geodesy, call require() with .js filenames. In
Node, require("./dms") and require("./dms.js") are the same path, but
in RequireJS's default configuration, the former is treated as a
module path and the latter is treated as a URL relative to the
document's base path (NOT the RequireJS baseUrl!

Enabling the Node support means that all modules need to be loaded
with module names and not relative URLs, so the plugin loader is
changed to give the module name of plugin javascript.